Add independent migrations config file
